busssiness update connected with backend

+++ lib/bussiness/domain/repository/i_bussiness_repository.dart
@@ -18,5 +18,5 @@ abstract class IBussinessRepository {
   
   Future<RatingInterceptor?> rateById(String bussinessId, int rating);
   Future<List<RatingInterceptor>?> getRates(String bussinessId);
-  Future<int?> getAverageRate(String bussinessId);
+  Future<double?> getAverageRate(String bussinessId);
 }
